ETDs: create local authorities for contributors
